backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
SQLAlchemy==2.0.29
asyncpg==0.29.0

# Redis (async) + RQ (sync queue)
redis==5.0.4
rq==1.15.1

# Env Loader
python-dotenv==1.0.1

# Utilities
requests==2.31.0
pydantic==2.7.0

# Async tools
aiofiles==23.2.1

# Optional AWS helpers
boto3==1.34.40
botocore==1.34.40
